素材管理结束

+++ src/views/article/article.vue
@@ -17,14 +17,15 @@
         </el-form-item>
 
         <el-form-item label="频道 :">
-          <el-select v-model="reqParams.channel_id" clearable placeholder="请选择">
+          <!-- <el-select v-model="reqParams.channel_id" clearable placeholder="请选择">
             <el-option
               v-for="item in channelOptions"
               :key="item.id"
               :label="item.name"
               :value="item.id"
             ></el-option>
-          </el-select>
+          </el-select> -->
+          <my-channel v-model="reqParams.channel_id"></my-channel>
         </el-form-item>
 
         <el-form-item label="日期 :">
@@ -117,16 +118,14 @@ export default {
     }
   },
   methods: {
-    async getChannelOptions () {
-      const {
-        data: { data }
-      } = await this.$axios.get('channels')
-      this.channelOptions = data.channels
-    },
+    // async getChannelOptions () {
+    //   const {
+    //     data: { data }
+    //   } = await this.$axios.get('channels')
+    //   this.channelOptions = data.channels
+    // },
     async getArticles () {
-      const {
-        data: { data }
-      } = await this.$axios.get('articles', { params: this.reqParams })
+      const { data: { data } } = await this.$axios.get('articles', { params: this.reqParams })
       this.articles = data.results
       this.total = data.total_count
     },
@@ -153,14 +152,25 @@ export default {
       // this.$router.push('/publish?id=' + id)
       this.$router.push({ path: 'publish', query: { id } })
     },
-    async del (id) {
-      await this.$axios.delete(`articles/${id}`)
-      this.$message.success('删除文章成功')
-      this.getArticles()
+    del (id) {
+      this.$confirm('此操作将永久删除该文件, 是否继续?', '提示', {
+        confirmButtonText: '确定',
+        cancelButtonText: '取消',
+        type: 'warning'
+      })
+        .then(async () => {
+          await this.$axios.delete(`articles/${id}`)
+          this.$message({
+            type: 'success',
+            message: '删除成功!'
+          })
+          this.getArticles()
+        })
+        .catch(() => {})
     }
   },
   created () {
-    this.getChannelOptions()
+    // this.getChannelOptions()
     this.getArticles()
   }
 }
